// src/models/TaskStateModel.ts
import type { TaskModel } from './TaskModel';

export type TaskStateModel = {
  tasks: TaskModel[];
  secondsRemaining: number;
  formattedSecondsRemaining: string;
  activeTask: TaskModel | null;
  currentCycle: number;
  config: {
    pomodori: number;
    pause: number;
    highPause: number;
  };
};